Introduce `Bot` interface
First step towards making the app logic generic across message
providers: implement a common interface for posting and updating
messages, replying in thread, healthchecking and doing user lookups.

Currently only slack and Discord implement the interface.
